Saving profile pictures in volume
Docker Compose:

Creation of a volume named profile-images
intended for storing profile pictures
uploaded by users.

Frontend:

We have a React component named
UserInformation that displays user
information, including their profile
picture.

To allow users to upload a new profile
picture, i added an <input> element of
type file to the component.

When a file is selected, we handle the
onChange event and send the file to the
backend using an HTTP request.

We use the axios library for the HTTP
request, which needed to be installed as a
dependency.

We updated the handleFileChange function
to send the file as multipart/form-data to
the /api/user/upload-profile-picture
endpoint.

Backend:

ProfilePictureController handles the
/api/user/upload-profile-picture endpoint.

The uploadProfilePicture method in the
controller receives a file using the
FileInterceptor from multer.

In the next comit the
'uploadProfilePicture' method will be
derived to the user.service.ts file.

Upon receiving the file, we generate a
unique filename and attempt to save it in
the profile-images volume.

Initially, i tried moving the file
directly using fs.renameSync(), but
encountered issues. Later, i used
fs.readFileSync() and fs.writeFileSync()
to read and write the file content
instead. After successfully saving the
profile picture, we delete the temporary
file and can proceed with any additional
logic (i still need to update the file
address in the database).

Some info about mutter and diskStorage:
1. Multer:

 * Multer is a middleware for
handling multipart/form-data, which is
commonly used for uploading files in HTML
forms.

 * It enables the server to
handle file uploads sent from the
client.

 * Multer works with the Express
framework and simplifies the process of
handling file uploads by providing an easy-
to-use API.

2. DiskStorage:

 * diskStorage is a storage engine provided
by Multer that allows you to define how
files should be stored on the disk.

 * It provides a way to specify the
destination directory and the filename of
the uploaded files.

 * With diskStorage, you can configure
the destination and  filename based on
your requirements.

* Together, multer and diskStorage
provide a convenient solution for
handling file uploads in Node.js
applications.

Multer handles the multipart/form-data
parsing, while diskStorage defines how
the uploaded  files should be stored on
the disk.

In our case we use diskStorage to specify
the destination directory as
'./uploads/profile-pictures' and generate
a unique filename for each uploaded file
using the current timestamp and a unique
identifier.